+++ src/app/feat-recruit/create_recruit/other_game_recruit.ts
@@ -12,20 +12,20 @@ import {
     Role,
 } from 'discord.js';
 
-import { ParticipantService } from '../../../../db/participant_service';
-import { RecruitService, RecruitType } from '../../../../db/recruit_service';
-import { UniqueRoleService } from '../../../../db/unique_role_service';
-import { log4js_obj } from '../../../../log4js_settings';
-import { searchDBMemberById } from '../../../common/manager/member_manager';
-import { searchMessageById } from '../../../common/manager/message_manager';
-import { assertExistCheck, exists, sleep } from '../../../common/others';
-import { RoleKeySet } from '../../../constant/role_key';
-import { sendErrorLogs } from '../../../logs/error/send_error_logs';
-import { embedRecruitDeleteButton, recruitActionRow } from '../../create_recruit/common/create_recruit_buttons';
-import { getVCReserveErrorMessage } from '../../create_recruit/condition_checks/vc_reserve_check';
-import { buildOtherGameRecruitText } from '../../create_recruit/common/recruit_text';
-import { createRecruitEvent } from '../../vc_reservation/recruit_event';
-import { sendRecruitSticky } from '../../sticky/recruit_sticky_messages';
+import { embedRecruitDeleteButton, recruitActionRow } from './common/create_recruit_buttons';
+import { buildOtherGameRecruitText } from './common/recruit_text';
+import { getVCReserveErrorMessage } from './condition_checks/vc_reserve_check';
+import { ParticipantService } from '../../../db/participant_service';
+import { RecruitService, RecruitType } from '../../../db/recruit_service';
+import { UniqueRoleService } from '../../../db/unique_role_service';
+import { log4js_obj } from '../../../log4js_settings';
+import { searchDBMemberById } from '../../common/manager/member_manager';
+import { searchMessageById } from '../../common/manager/message_manager';
+import { assertExistCheck, exists, sleep } from '../../common/others';
+import { RoleKeySet } from '../../constant/role_key';
+import { sendErrorLogs } from '../../logs/error/send_error_logs';
+import { sendRecruitSticky } from '../sticky/recruit_sticky_messages';
+import { createRecruitEvent } from '../vc_reservation/recruit_event';
 
 const logger = log4js_obj.getLogger('recruit');
