+++ src/app/handlers/command_handler.ts
@@ -14,14 +14,14 @@ import { shutdown } from '../feat-admin/shutdown/shutdown_process';
 import { uniqueChannelSettingsHandler } from '../feat-admin/unique_channel_settings/unique_channel_settings_hanlder.js';
 import { uniqueRoleSettingsHandler } from '../feat-admin/unique_role_settings/unique_role_settings_hanlder.js';
 import { anarchyRecruit } from '../feat-recruit/create_recruit/anarchy_recruit.js';
+import { buttonRecruit } from '../feat-recruit/create_recruit/button_recruit.js';
 import { eventRecruit } from '../feat-recruit/create_recruit/event_recruit.js';
 import { festRecruit } from '../feat-recruit/create_recruit/fest_recruit.js';
+import { otherGameRecruit } from '../feat-recruit/create_recruit/other_game_recruit';
+import { privateRecruit } from '../feat-recruit/create_recruit/private_recruit';
 import { regularRecruit } from '../feat-recruit/create_recruit/regular_recruit.js';
 import { salmonRecruit } from '../feat-recruit/create_recruit/salmon_recruit.js';
-import { buttonRecruit } from '../feat-recruit/interactions/commands/button_recruit.js';
 import { closeCommand } from '../feat-recruit/interactions/commands/close.js';
-import { otherGameRecruit } from '../feat-recruit/interactions/commands/other_game_recruit';
-import { privateRecruit } from '../feat-recruit/interactions/commands/private_recruit';
 import { handleIkabuExperience } from '../feat-utils/other/experience.js';
 import { handleFriendCode } from '../feat-utils/other/friendcode';
 import { handleHelp } from '../feat-utils/other/help.js';
